Refactoring: Restructure
